refactor: move Loom types to a new file and update imports
- Created a new `types.ts` file to define `LoomContext`, `LoomConfig`, and `LoomInstance` interfaces.
- Updated `index.ts` and `loom.ts` to import types from the new file, improving organization and clarity.
