+++ src/index.ts
@@ -2,4 +2,10 @@ export { sequences } from "./generators/sequences";
 export { zwjSequences } from "./generators/zwj-sequences";
 
 export { createLoom } from "./loom";
-export type { LoomConfig, LoomContext, LoomInstance } from "./loom";
+export type {
+  LoomConfig,
+  LoomContext,
+  LoomContextHelperFn,
+  LoomContextHelpers,
+  LoomInstance,
+} from "./types";
